requirements.txt
streamlit>=1.38
folium>=0.17
streamlit-folium>=0.22
pandas>=2.2
pyyaml>=6.0.2
